+++ libraries/microsoft-agents-testing/microsoft_agents/testing/_defaults.py
@@ -1,0 +1,15 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+_SERVICE_URL = "https://test.com"
+
+_CONV_ID = "convo1"
+_CONV_NAME = "Conversation 1"
+
+_BOT_ID = "bot"
+_BOT_NAME = "Bot"
+
+_USER_ID = "user1"
+_USER_NAME = "User 1"
+
+_LOCALE = "en-US"